backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15,<3.17
django-cors-headers>=4.6,<4.8
channels>=4.1,<4.3
daphne>=4.1,<4.3
numpy>=2.2,<2.4
